requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
alembic==1.12.1
authlib==1.2.1
click==8.1.7
pandas==2.1.4
pdfkit==1.0.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
openpyxl==3.1.2
Flask-CORS
gunicorn